+++ apps/prairielearn/src/lib/server-jobs.ts
@@ -1,14 +1,18 @@
 import AnsiUp from 'ansi_up';
 import execa = require('execa');
 import { z } from 'zod';
+import { setTimeout as sleep } from 'node:timers/promises';
+import * as _ from 'lodash';
+
 import * as Sentry from '@prairielearn/sentry';
 import { logger } from '@prairielearn/logger';
 import { loadSqlEquiv, queryAsync, queryRow, queryRows } from '@prairielearn/postgres';
+import { checkSignedToken, generateSignedToken } from '@prairielearn/signed-token';
 
 import { chalk, chalkDim } from './chalk';
-import * as serverJobs from './server-jobs-legacy';
 import * as socketServer from './socket-server';
-import { Job, JobSchema } from './db-types';
+import { IdSchema, Job, JobSchema, JobSequenceSchema, UserSchema } from './db-types';
+import { config } from './config';
 
 const sql = loadSqlEquiv(__filename);
 
@@ -46,6 +50,33 @@ export interface ServerJobExecutor {
 }
 
 export type ServerJobExecutionFunction = (job: ServerJob) => Promise<void>;
+
+const JobSequenceWithJobsSchema = JobSequenceSchema.extend({
+  start_date_formatted: z.string().nullable(),
+  finish_date_formatted: z.string().nullable(),
+  user_uid: UserSchema.shape.uid.nullable(),
+  authn_user_uid: UserSchema.shape.uid.nullable(),
+  job_count: z.coerce.number(),
+  jobs: z.array(
+    JobSchema.extend({
+      start_date_formatted: z.string().nullable(),
+      finish_date_formatted: z.string().nullable(),
+      user_uid: UserSchema.shape.uid.nullable(),
+      authn_user_uid: UserSchema.shape.uid.nullable(),
+    }),
+  ),
+});
+
+/**
+ * Store currently active job information in memory. This is used
+ * to accumulate stderr/stdout, which are only written to the DB
+ * once the job is finished.
+ */
+export const liveJobs: Record<string, ServerJobImpl> = {};
+
+/********************************************************************/
+
+let heartbeatIntervalId: ReturnType<typeof setInterval> | null = null;
 
 /**
  * Internal error subclass so we can identify when `fail()` is called.
@@ -213,7 +244,7 @@ class ServerJobImpl implements ServerJob, ServerJobExecutor {
       }
     }
 
-    delete serverJobs.liveJobs[this.jobId];
+    delete liveJobs[this.jobId];
 
     await queryAsync(sql.update_job_on_finish, {
       job_sequence_id: this.jobSequenceId,
@@ -252,10 +283,209 @@ export async function createServerJob(options: CreateServerJobOptions): Promise<
   );
 
   const serverJob = new ServerJobImpl(job_sequence_id, job_id);
-  serverJobs.liveJobs[job_id] = serverJob;
+  liveJobs[job_id] = serverJob;
   return serverJob;
 }
 
 export async function selectJobsByJobSequenceId(jobSequenceId: string): Promise<Job[]> {
   return await queryRows(sql.select_job_output, { job_sequence_id: jobSequenceId }, JobSchema);
+}
+
+/*
+  SocketIO configuration
+
+  ##########################################
+  Room 'job-231' for job_id = 231 in DB.
+
+  Receives messages:
+  'joinJob', arguments: job_id, returns: status, output
+
+  Sends messages:
+  'change:output', arguments: job_id, output
+  'update', arguments: job_id
+
+  ##########################################
+  Room 'jobSequence-593' for job_sequence_id = 593 in DB.
+
+  Receives messages:
+  'joinJobSequence', arguments: job_sequence_id, returns: job_count
+
+  Sends messages:
+  'update', arguments: job_sequence_id
+
+  ##########################################
+
+  'update' events are sent for bulk changes to the object when
+  there is no specific 'change' event available.
+
+  For example, an 'update' will not be sent when output changes
+  because the more specific 'change:output' event will be sent
+  instead.
+*/
+
+export function init() {
+  socketServer.io.on('connection', connection);
+
+  // Start a periodic task to heartbeat all live jobs. We don't use a cronjob
+  // for this because we want this to run for every host.
+  heartbeatIntervalId = setInterval(() => {
+    const jobIds = Object.keys(liveJobs);
+    if (jobIds.length === 0) return;
+
+    queryAsync(sql.update_heartbeats, { job_ids: jobIds }).catch((err) => {
+      logger.error('Error updating heartbeats for live server jobs', err);
+    });
+  }, config.serverJobHeartbeatIntervalSec * 1000);
+}
+
+export async function stop() {
+  // Wait until all jobs have finished.
+  while (Object.keys(liveJobs).length > 0) {
+    await sleep(100);
+  }
+
+  // Only once all jobs are finished should we stop sending heartbeats.
+  if (heartbeatIntervalId) {
+    clearInterval(heartbeatIntervalId);
+    heartbeatIntervalId = null;
+  }
+}
+
+export function connection(socket) {
+  socket.on('joinJob', function (msg, callback) {
+    if (!_.has(msg, 'job_id')) {
+      logger.error('socket.io joinJob called without job_id');
+      return;
+    }
+
+    // Check authorization of the requester.
+    if (!checkSignedToken(msg.token, { jobId: msg.job_id.toString() }, config.secretKey)) {
+      logger.error(`joinJob called with invalid token for job_id ${msg.job_id}`);
+      return;
+    }
+
+    socket.join('job-' + msg.job_id);
+    queryRow(sql.select_job, { job_id: msg.job_id }, JobSchema).then(
+      (job) => {
+        const status = job.status;
+
+        const ansiUp = new AnsiUp();
+        const output = ansiUp.ansi_to_html(liveJobs[msg.job_id]?.output ?? job.output ?? '');
+
+        callback({ status, output });
+      },
+      (err) => {
+        logger.error('socket.io joinJob error selecting job_id ' + msg.job_id, err);
+      },
+    );
+  });
+
+  socket.on('joinJobSequence', function (msg, callback) {
+    if (!_.has(msg, 'job_sequence_id')) {
+      logger.error('socket.io joinJobSequence called without job_sequence_id');
+      return;
+    }
+
+    // Check authorization of the requester.
+    if (
+      !checkSignedToken(
+        msg.token,
+        { jobSequenceId: msg.job_sequence_id.toString() },
+        config.secretKey,
+      )
+    ) {
+      logger.error(
+        `joinJobSequence called with invalid token for job_sequence_id ${msg.job_sequence_id}`,
+      );
+      return;
+    }
+
+    socket.join('jobSequence-' + msg.job_id);
+    queryRow(
+      sql.select_job_sequence,
+      { job_sequence_id: msg.job_sequence_id },
+      JobSequenceSchema.extend({ job_count: z.coerce.number() }),
+    ).then(
+      ({ job_count }) => {
+        callback({ job_count });
+      },
+      (err) => {
+        logger.error(
+          'socket.io joinJobSequence error selecting job_sequence_id ' + msg.job_sequence_id,
+          err,
+        );
+      },
+    );
+  });
+}
+
+export async function errorAbandonedJobs() {
+  const abandonedJobs = await queryRows(
+    sql.select_abandoned_jobs,
+    { timeout_secs: config.serverJobsAbandonedTimeoutSec },
+    z.object({ id: IdSchema, job_sequence_id: IdSchema.nullable() }),
+  );
+
+  for (const row of abandonedJobs) {
+    logger.debug('Job abandoned by server, id: ' + row.id);
+    try {
+      await queryAsync(sql.update_job_on_error, {
+        job_id: row.id,
+        output: null,
+        error_message: 'Job abandoned by server',
+      });
+    } catch (err) {
+      logger.error('errorAbandonedJobs: error updating job on error', err);
+    } finally {
+      socketServer.io.to('job-' + row.id).emit('update');
+      if (row.job_sequence_id != null) {
+        socketServer.io.to('jobSequence-' + row.job_sequence_id).emit('update');
+      }
+    }
+  }
+
+  const abandonedJobSequences = await queryRows(sql.error_abandoned_job_sequences, IdSchema);
+  abandonedJobSequences.forEach(function (job_sequence_id) {
+    socketServer.io.to('jobSequence-' + job_sequence_id).emit('update');
+  });
+}
+
+export async function getJobSequence(job_sequence_id: string, course_id: string | null) {
+  const jobSequence = await queryRow(
+    sql.select_job_sequence_with_course_id_as_json,
+    { job_sequence_id, course_id },
+    JobSequenceWithJobsSchema,
+  );
+
+  // Generate a token to authorize websocket requests for this job sequence.
+  const jobSequenceTokenData = { jobSequenceId: job_sequence_id.toString() };
+  return {
+    ...jobSequence,
+    token: generateSignedToken(jobSequenceTokenData, config.secretKey),
+    jobs: jobSequence.jobs.map((job) => {
+      // Also generate a token for each job.
+      const jobTokenData = { jobId: job.id.toString() };
+      return { ...job, token: generateSignedToken(jobTokenData, config.secretKey) };
+    }),
+  };
+}
+
+/**
+ * Resolves with a job sequence, where each job's output has been turned into
+ * markup with `ansi_up`.
+ */
+export async function getJobSequenceWithFormattedOutput(
+  job_sequence_id: string,
+  course_id: string | null,
+) {
+  const jobSequence = await getJobSequence(job_sequence_id, course_id);
+  const ansiup = new AnsiUp();
+  return {
+    ...jobSequence,
+    jobs: jobSequence.jobs.map((job) => ({
+      ...job,
+      output_raw: job.output,
+      output: job.output ? ansiup.ansi_to_html(job.output) : '',
+    })),
+  };
 }
